Extract lifetime.rs inline tests into lifetime/tests.rs
Mechanical move of the ~616-line #[cfg(test)] module; all production
code stays in src/player/lifetime.rs, where check-architecture.sh pins
the orphan-recovery call ordering by exact string. The size-baseline
row for lifetime.rs drops 1591 -> 975 so the ratchet locks in the
reduction (only that row is touched).
